rearranged code for cohort app
