+++ exports/email_templates/confirmation_de_requete.html
@@ -3,10 +3,11 @@ Nous vous confirmons la réception de votre demande d'export de données concern
 <br>
 Un lien de téléchargement vous sera envoyé par email lorsque les fichiers seront disponibles.
 <br>
-Un délai de 24 h maximum peut s'appliquer en fonction du volume des données.
+Un délai de 24h maximum peut s'appliquer en fonction du volume des données.
 <br>
 <br>
-Les fichiers de données seront alors disponibles et téléchargeables uniquement par vous pendant 7 jours. Passé cette date, le téléchargement ne sera plus possible.
+Les fichiers de données seront alors disponibles et téléchargeables uniquement par vous pendant 7 jours.
+Passé cette date, le téléchargement ne sera plus possible.
 <br>
 <br>
 Pour toute question, contactez le support Cohort360 à KEY_CONTACT_MAIL
