+++ exports/email_templates/confirmation_de_requete.txt
@@ -1,7 +1,7 @@
 Nous vous confirmons la réception de votre demande d'export de données concernant la cohorte numéro KEY_COHORT_ID.
 
 Un lien de téléchargement vous sera envoyé par email lorsque les fichiers seront disponibles.
-Un délai de 24 h maximum peut s'appliquer en fonction du volume des données.
+Un délai de 24h maximum peut s'appliquer en fonction du volume des données.
 
 Les fichiers de données seront alors disponibles et téléchargeables uniquement par vous pendant 7 jours. Passé cette date, le téléchargement ne sera plus possible.
