+++ exports/email_templates/confirmation_de_requete_hive.html
@@ -3,7 +3,7 @@ Nous vous confirmons la réception de votre demande d'export de données vers Hi
 <br>
 Vous serez informé.e par email lorsque les données seront disponibles dans votre espace Jupyter.
 <br>
-Un délai de 24 h maximum peut s'appliquer en fonction du volume des données.
+Un délai de 24h maximum peut s'appliquer en fonction du volume des données.
 <br>
 <br>
 Pour toute question, contactez le support Cohort360 à KEY_CONTACT_MAIL
