+++ exports/email_templates/resultat_requete_echec.txt
@@ -1,4 +1,4 @@
-Malheureusement, votre demande d'export de données concernant la cohorte numéro KEY_COHORT_ID n'a pas pu aboutir, pour la raison suivante :
+Malheureusement, votre demande d'export de données concernant la cohorte numéro KEY_COHORT_ID n'a pas pu aboutir, pour la raison suivante:
 
 KEY_ERROR_MESSAGE
