+++ exports/email_templates/resultat_requete_succes.html
@@ -1,7 +1,7 @@
 Votre export de données concernant la cohorte numéro KEY_COHORT_ID a été généré avec succès.
 <br>
 <br>
-Téléchargez les fichiers en suivant le lien ci-dessous en utilisant votre code APH et mot de passe de session Windows :
+Téléchargez les fichiers en suivant le lien ci-dessous en utilisant votre code APH et mot de passe de session Windows:
 <br>
 <br>
 KEY_DOWNLOAD_URL
