+++ exports/email_templates/resultat_requete_succes_hive.txt
@@ -2,7 +2,7 @@ Votre export de données concernant la cohorte numéro KEY_COHORT_ID a été gé
 
 Les données sont accessibles dans votre espace Jupyter dans la base de données nommée KEY_DATABASE_NAME.
 
-Pour une première utilisation, vous trouverez à ce lien une aide pour l'ouverture des fichiers :
+Pour une première utilisation, vous trouverez à ce lien une aide pour l'ouverture des fichiers:
 https://gitlab.eds.aphp.fr/public-resources/eds-doc/-/wikis/Export
 
 Pour toute question, contactez le support Cohort360 à KEY_CONTACT_MAIL
