+++ exports/views/__init__.py
@@ -1,4 +1,2 @@
 from .cohort import CohortViewSet
 from .export_request import ExportRequestViewSet
-
-__all__ = ["ExportRequestViewSet", "CohortViewSet"]
