+++ workspaces/views/__init__.py
@@ -4,6 +4,3 @@ from .kernel import KernelViewSet
 from .ldap_group import LdapGroupViewSet
 from .project import ProjectViewSet
 from .ranger_hive_policy import RangerHivePolicyViewSet
-
-__all__ = ["AccountViewSet", "JupyterMachineViewSet", "KernelViewSet",
-           "LdapGroupViewSet", "ProjectViewSet", "RangerHivePolicyViewSet"]
